pyproject.toml
[project]
name = "molly-tagger"
version = "0.1.0"
description = "LLM-based tag extraction for Obsidian notes"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Tsinglung Tseng"}]
requires-python = ">=3.10"
keywords = ["obsidian", "nlp", "ner", "entity-extraction", "llm", "tagging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "pyyaml>=6.0",
    "watchdog>=4.0.0",
    "python-frontmatter>=1.0.0",
    "requests>=2.31.0",
]

[project.urls]
Repository = "https://github.com/Tsinglung-Tseng/molly.tagger"
